# Revision performance improvements