### Description

Builds on the changes in #10577, but for revision creation.

When a revision is being created, if it contains any Matrix blocks, those blocks are no longer directly duplicated for the revision. Instead, we create _revisions_ for each of the Matrix blocks, which are assigned to the owner’s revision.

Revision creation still duplicates the element under the hood, so the initial revision creation isn’t much different from the normal block duplication we’ve done previously. _However_, revisions are only ever created for elements that have been edited since the last time a revision was created. So in effect, Matrix blocks will only ever be duplicated for an owner’s new revision if they have actually been edited. 

Here’s a look at the performance gains when saving an entry with 100 blocks and 10 sites:

<img width="748" alt="Three bar charts showing the performance gains that this PR introduces. See the table below for the raw data" src="https://user-images.githubusercontent.com/47792/154358361-26fe0611-02e8-4046-97d9-33680d0db99a.png">

|  | Before | After |
| -- | -- | -- |
| **Time** | 41.36s | 1.6s |
| **Queries** | 13,090 | 453 |
| **Memory** | 108MB | 38MB |

### Related issues

- #9802